+++ src/main/java/com/unascribed/yttr/content/block/mechanism/FlopperBlockEntity.java
@@ -2,7 +2,6 @@ package com.unascribed.yttr.content.block.mechanism;
 
 import com.unascribed.yttr.init.YBlockEntities;
 import com.unascribed.yttr.init.YBlocks;
-import com.unascribed.yttr.mixin.accessor.AccessorBlockEntity;
 import com.unascribed.yttr.util.YTickable;
 
 import net.minecraft.block.BlockState;
@@ -16,26 +15,13 @@ import net.minecraft.util.shape.VoxelShape;
 
 public class FlopperBlockEntity extends HopperBlockEntity implements YTickable {
 
-	private BlockState realState;
-
 	public FlopperBlockEntity(BlockPos pos, BlockState state) {
 		super(pos, state);
 	}
 	
 	@Override
 	public void tick() {
-		realState = getCachedState();
-		try {
-			((AccessorBlockEntity)this).yttr$setCachedState(realState.with(FlopperBlock.FACING, Direction.DOWN));
-			HopperBlockEntity.serverTick(world, pos, realState, this);
-		} finally {
-			((AccessorBlockEntity)this).yttr$setCachedState(realState);
-			realState = null;
-		}
-	}
-	
-	public BlockState getRealState() {
-		return realState == null ? getCachedState() : realState;
+		HopperBlockEntity.serverTick(world, pos, getCachedState().with(FlopperBlock.FACING, Direction.DOWN), this);
 	}
 	
 	@Override
@@ -50,7 +36,7 @@ public class FlopperBlockEntity extends HopperBlockEntity implements YTickable {
 	
 	@Override
 	public VoxelShape getInputAreaShape() {
-		return YBlocks.FLOPPER.getFunnelShape(getRealState());
+		return YBlocks.FLOPPER.getFunnelShape(getCachedState());
 	}
 	
 }