+++ src/main/java/com/unascribed/yttr/content/block/mechanism/LevitationChamberBlock.java
@@ -3,6 +3,7 @@ package com.unascribed.yttr.content.block.mechanism;
 import java.util.Locale;
 
 import com.unascribed.yttr.init.YBlocks;
+import com.unascribed.yttr.util.YTickable;
 
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockEntityProvider;
@@ -10,6 +11,8 @@ import net.minecraft.block.BlockState;
 import net.minecraft.block.Blocks;
 import net.minecraft.block.ShapeContext;
 import net.minecraft.block.entity.BlockEntity;
+import net.minecraft.block.entity.BlockEntityTicker;
+import net.minecraft.block.entity.BlockEntityType;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.ItemPlacementContext;
 import net.minecraft.state.StateManager.Builder;
@@ -99,6 +102,11 @@ public class LevitationChamberBlock extends Block implements BlockEntityProvider
 	@Override
 	public VoxelShape getOutlineShape(BlockState state, BlockView world, BlockPos pos, ShapeContext context) {
 		return SHAPE;
+	}
+	
+	@Override
+	public <T extends BlockEntity> BlockEntityTicker<T> getTicker(World world, BlockState state, BlockEntityType<T> type) {
+		return YTickable::tick;
 	}
 
 }