+++ src/main/java/com/unascribed/yttr/init/YHandledScreens.java
@@ -5,6 +5,7 @@ import static java.lang.annotation.RetentionPolicy.RUNTIME;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
+import java.util.function.Consumer;
 
 import com.unascribed.yttr.Yttr;
 import com.unascribed.yttr.client.screen.handled.AmmoPackScreen;
@@ -15,6 +16,7 @@ import com.unascribed.yttr.client.screen.handled.InRedOscillatorScreen;
 import com.unascribed.yttr.client.screen.handled.MagtankScreen;
 import com.unascribed.yttr.client.screen.handled.ProjectTableScreen;
 import com.unascribed.yttr.client.screen.handled.RafterScreen;
+import com.unascribed.yttr.client.screen.handled.SSDScreen;
 import com.unascribed.yttr.client.screen.handled.SuitStationScreen;
 import com.unascribed.yttr.client.screen.handled.VoidFilterScreen;
 import com.unascribed.yttr.inventory.AmmoPackScreenHandler;
@@ -25,42 +27,61 @@ import com.unascribed.yttr.inventory.InRedOscillatorScreenHandler;
 import com.unascribed.yttr.inventory.MagtankScreenHandler;
 import com.unascribed.yttr.inventory.ProjectTableScreenHandler;
 import com.unascribed.yttr.inventory.RafterScreenHandler;
+import com.unascribed.yttr.inventory.SSDScreenHandler;
 import com.unascribed.yttr.inventory.SuitStationScreenHandler;
 import com.unascribed.yttr.inventory.VoidFilterScreenHandler;
 
 import net.minecraft.client.gui.screen.ingame.HandledScreen;
+import net.minecraft.entity.player.PlayerInventory;
 import net.minecraft.screen.ScreenHandlerType;
+import net.minecraft.screen.slot.Slot;
 import net.minecraft.util.registry.Registry;
 
 public class YHandledScreens {
 
 	@Screen(CentrifugeScreen.class)
-	public static final ScreenHandlerType<CentrifugeScreenHandler> CENTRIFUGE = Registry.register(Registry.SCREEN_HANDLER, Yttr.id("centrifuge"), new ScreenHandlerType<>(CentrifugeScreenHandler::new));
+	public static final ScreenHandlerType<CentrifugeScreenHandler> CENTRIFUGE = new ScreenHandlerType<>(CentrifugeScreenHandler::new);
 	@Screen(SuitStationScreen.class)
-	public static final ScreenHandlerType<SuitStationScreenHandler> SUIT_STATION = Registry.register(Registry.SCREEN_HANDLER, Yttr.id("suit_station"), new ScreenHandlerType<>(SuitStationScreenHandler::new));
+	public static final ScreenHandlerType<SuitStationScreenHandler> SUIT_STATION = new ScreenHandlerType<>(SuitStationScreenHandler::new);
 	@Screen(VoidFilterScreen.class)
-	public static final ScreenHandlerType<VoidFilterScreenHandler> VOID_FILTER = Registry.register(Registry.SCREEN_HANDLER, Yttr.id("void_filter"), new ScreenHandlerType<>(VoidFilterScreenHandler::new));
+	public static final ScreenHandlerType<VoidFilterScreenHandler> VOID_FILTER = new ScreenHandlerType<>(VoidFilterScreenHandler::new);
 	@Screen(DSUScreen.class)
-	public static final ScreenHandlerType<DSUScreenHandler> DSU = Registry.register(Registry.SCREEN_HANDLER, Yttr.id("dsu"), new ScreenHandlerType<>(DSUScreenHandler::new));
+	public static final ScreenHandlerType<DSUScreenHandler> DSU = new ScreenHandlerType<>(DSUScreenHandler::new);
 	@Screen(MagtankScreen.class)
-	public static final ScreenHandlerType<MagtankScreenHandler> MAGTANK = Registry.register(Registry.SCREEN_HANDLER, Yttr.id("magtank"), new ScreenHandlerType<>(MagtankScreenHandler::new));
+	public static final ScreenHandlerType<MagtankScreenHandler> MAGTANK = new ScreenHandlerType<>(MagtankScreenHandler::new);
 	@Screen(CanFillerScreen.class)
-	public static final ScreenHandlerType<CanFillerScreenHandler> CAN_FILLER = Registry.register(Registry.SCREEN_HANDLER, Yttr.id("can_filler"), new ScreenHandlerType<>(CanFillerScreenHandler::new));
+	public static final ScreenHandlerType<CanFillerScreenHandler> CAN_FILLER = new ScreenHandlerType<>(CanFillerScreenHandler::new);
 	@Screen(AmmoPackScreen.class)
-	public static final ScreenHandlerType<AmmoPackScreenHandler> AMMO_PACK = Registry.register(Registry.SCREEN_HANDLER, Yttr.id("ammo_pack"), new ScreenHandlerType<>(AmmoPackScreenHandler::new));
+	public static final ScreenHandlerType<AmmoPackScreenHandler> AMMO_PACK = new ScreenHandlerType<>(AmmoPackScreenHandler::new);
 	@Screen(InRedOscillatorScreen.class)
-	public static final ScreenHandlerType<InRedOscillatorScreenHandler> INRED_OSCILLATOR = Registry.register(Registry.SCREEN_HANDLER, Yttr.id("inred_oscillator"), new ScreenHandlerType<>(InRedOscillatorScreenHandler::new));
+	public static final ScreenHandlerType<InRedOscillatorScreenHandler> INRED_OSCILLATOR = new ScreenHandlerType<>(InRedOscillatorScreenHandler::new);
 	@Screen(RafterScreen.class)
-	public static final ScreenHandlerType<RafterScreenHandler> RAFTING = Registry.register(Registry.SCREEN_HANDLER, Yttr.id("rafter"), new ScreenHandlerType<>(RafterScreenHandler::new));
+	public static final ScreenHandlerType<RafterScreenHandler> RAFTING = new ScreenHandlerType<>(RafterScreenHandler::new);
 	@Screen(ProjectTableScreen.class)
-	public static final ScreenHandlerType<ProjectTableScreenHandler> PROJECT_TABLE = Registry.register(Registry.SCREEN_HANDLER, Yttr.id("project_table"), new ScreenHandlerType<>(ProjectTableScreenHandler::new));
+	public static final ScreenHandlerType<ProjectTableScreenHandler> PROJECT_TABLE = new ScreenHandlerType<>(ProjectTableScreenHandler::new);
+	@Screen(SSDScreen.class)
+	public static final ScreenHandlerType<SSDScreenHandler> SSD = new ScreenHandlerType<>(SSDScreenHandler::new);
 	
-	public static void init() {}
+	public static void init() {
+		Yttr.autoRegister(Registry.SCREEN_HANDLER, YHandledScreens.class, ScreenHandlerType.class);
+	}
 	
 	@Retention(RUNTIME)
 	@Target(FIELD)
 	public @interface Screen {
 		Class<? extends HandledScreen<?>> value();
+	}
+
+	public static void addPlayerSlots(Consumer<Slot> h, PlayerInventory playerInv, int oX, int oY) {
+		for (int y = 0; y < 3; y++) {
+			for (int x = 0; x < 9; x++) {
+				h.accept(new Slot(playerInv, x + y * 9 + 9, oX + (x * 18), oY + (y * 18)));
+			}
+		}
+
+		for (int i = 0; i < 9; i++) {
+			h.accept(new Slot(playerInv, i, oX + (i * 18), oY + 58));
+		}
 	}
 	
 }