+++ src/main/java/com/unascribed/yttr/inventory/CanFillerScreenHandler.java
@@ -56,15 +56,7 @@ public class CanFillerScreenHandler extends ScreenHandler {
 		addSlot(new CFSlot(canFiller, 3, 71, 86, is -> false));
 		addSlot(new CFSlot(canFiller, 4, 89, 86, is -> false));
 		
-		for (int y = 0; y < 3; ++y) {
-			for (int x = 0; x < 9; ++x) {
-				addSlot(new Slot(playerInv, x + y * 9 + 9, 8 + x * 18, 119 + y * 18));
-			}
-		}
-
-		for (int i = 0; i < 9; ++i) {
-			addSlot(new Slot(playerInv, i, 8 + i * 18, 177));
-		}
+		YHandledScreens.addPlayerSlots(this::addSlot, playerInv, 8, 119);
 		
 		addProperties(properties);
 	}