+++ src/main/java/com/unascribed/yttr/inventory/ProjectTableScreenHandler.java
@@ -118,16 +118,8 @@ public class ProjectTableScreenHandler extends AbstractRecipeScreenHandler<Craft
 				addSlot(new Slot(inv, x + y * 9 + 9, 8 + x * 18, 84 + y * 18));
 			}
 		}
-
-		for (int y = 0; y < 3; y++) {
-			for (int x = 0; x < 9; x++) {
-				addSlot(new Slot(playerInventory, x + y * 9 + 9, 8 + x * 18, 133 + y * 18));
-			}
-		}
-
-		for (int i = 0; i < 9; i++) {
-			addSlot(new Slot(playerInventory, i, 8 + i * 18, 191));
-		}
+		
+		YHandledScreens.addPlayerSlots(this::addSlot, playerInventory, 8, 133);
 		
 		updateResult(syncId, player.world, player, input, result, this);