+++ CHANGELOG.md
@@ -1,2 +1,6 @@
 * Fixed crash when shooting ghast fireballs with the rifle
 * Fixed durability bars not rendering properly on ultrapure tools (thanks, Noobulus)
+* Void Bucket now has a trash icon and tooltips warning you that it will destroy things
+* Void Bucket now only voids items on right-click instead of both left and right
+* Fixed the Flopper, Dopper, and Levitation Chamber being in various states of not working
+* Added the Scalable Storage Drum or SSD, an inventory that can be adjusted from 1 to 8 slots big